Fix file naming in the player module
And this is why all changes should go through review.
